[Java.Interop] Implement JniEnvironment.GetExceptionForLastThrowable()
I was getting a random SIGSEGV/etc. because I was looking for a method
that doesn't exist, which _should_ throw, but because
JniEnvironment.GetExceptionForLastThrowable() didn't do anything
execution continued, and sadness ensued.

Implement JniEnvironment.GetExceptionForLastThrowable(). It currently
just creates a new JniException instance with the result of
java.lang.Object.toString() from the Java exception object.