[Java.Interop] Add JniType.GetCached*() methods.
The intent of the JniType.GetCacched*() methods is twofold:

 1. For "delayed lookup" semantics in a code generator.
 2. To support ID invalidation (see JavaVM.ClearTrackedReferences()
    and 5009e28c).

As part of this, a "class binding" would be akin to the new
TestObjectBinding.cs file [0]. This allows types and members to be
looked in an on-demand, thread-safe fashion, looking the member up
only once (reducing JNI invocations).

Additionally, JniStrings is made public (as create a new JniString
class with instance members seems overkill), and the
JniStrings.ToString() method is altered to take a (new)
JniHandleOwnership parameter so that the caller can pass "ownership"
(and disposal) of the handle to JniStrings.ToString().

Interesting side-note: because of the strongly-typed nature of all our
SafeHandle subclasses, we don't need
JniHandleOwnership.TransferGlobalRef or .TransferLocalRef as monodroid
requires; we can just have a single JniHandleOwnership.Transfer value.

[0]:

	partial class TestObjectBinding
	{
		static JniType _TypeRef;
		static JniType TypeRef {
			get {return JniType.GetCachedJniType (ref _TypeRef, "java/lang/Object");}
		}

		public JniGlobalReference SafeHandle {get; private set;}

		JniInstanceMethodID Object_ctor;
		public TestObjectBinding ()
		{
			TypeRef.GetCachedConstructor (ref Object_ctor, "()V");
			using (var lref = TypeRef.NewObject (Object_ctor))
				SafeHandle = lref.NewGlobalRef ();
		}

		JniInstanceMethodID Object_toString;
		public override string ToString ()
		{
			return JniStrings.ToString (
					TypeRef.GetCachedInstanceMethod (ref Object_toString, "toString", "()Ljava/lang/String;")
						.InvokeObjectMethod (SafeHandle),
					JniHandleOwnership.Transfer);
		}
	}