[Java.Interop] Add Java.Interop.JavaObject.
The beginnings of cross-VM object-reference support, as eluded to in
commit 8d9b11e9. From README.md:

> Types with a `Java` prefix are "high-level" types which participate in cross-VM
> object-reference semantics, e.g. you could add a `JavaObject` subclass to a
> Java-side collection, perform a GC, and the instance will survive the GC.

Java.Interop.JavaObject is a (bad) first attempt at doing that: it has
a finalizer which will convert the Global reference into a Weak Global
reference, perform a Java-side GC, then create a Global reference from
the Weak global reference. If the returned Global referenceis valid,
then something in "Java land" is referencing the Java-side instance,
and the instance shouldn't be collected, so we use
GC.ReRegisterForFinalize() to keep the JavaObject instance alive.

The problem with the above approach is that it only deals with single
instances, not object graphs, so a Java-side array which references
a JavaObject which reference the same array would result in an
uncollectable object graph (as the JavaObject finalizer wouldn't
change the referenced array reference, so the array reference would
keep the JavaObject alive).

The proper solution here is to use Mono's GC bridge, which will be
done in a future commit. I don't know how to support .NET, or if it's
even worth worrying about how to support .NET at this time.

"Along for the ride" of the overarching theme of "support JavaObject"
is a ton of other "stuff":

  * A partial internal binding of java.lang.Object,
    java.lang.Runtime(), and java.lang.System;
  * The addition of JniReferenceSafeHandle.Null
    (apparently passing `null` for JniReferenceSafeHandle parameters
    displeases mono and results in a crash).
  * The introduction of Java.Interop.IJavaObject (why not?)
  * Making JniArrays public (a short-term fix to make the unit tests
    for JavaObject sane; this will be "fixed" later).
  * Addition of the JniInvocationHandle type (to permit invoking
    "arbitrary" IntPtr handle references;
    JniInvocationHandle.Dispose() does nothing).
  * IJavaObject reference tracking within JavaVM, so that we can
    obtain the IJavaObject instance that corresponds to a given JNI
    handle.
  * The addition of JavaVM.GetObject(), a wrapper over the above
    IJavaObject tracking.