[Java.Interop] Sanitize instance tracking, JniEnvironment disposal
(This is way too big, but not big enough to do the work of splitting
this up into smaller patches...)

The "problem" was that JniEnvironment wouldn't remove itself from
JavaVM when it was disposed.

This in turn dovetailed with the observation that JavaVM.Environments
and JavaVM.TrackedInstances were nearly identical.

Thus: Remove JavaVM.Environments and use JavaVM.TrackedInstances
instead.

Then we hit a semantic problem: some things we need to track
(JniEnvironment, JniType), some things are optional (JniFieldID,
JniMethodID), and both were using JavaVM.Track(), which was confusing.

Rename the "optional track" method to JavaVM.TrackID(), which
dovetails nicely with the existing JavaVM.TrackIDs property.

Finally:
  - JniEnvironment.Object_class needs to use java.lang.Object, not
    java.lang.Class. (Doh!)
  - Rename JniEnvironment.GetEnvironmentFromHandle() to
    JniEnvironment.CheckCurrent().
  - Fix/sanitize JniEnvironment.CheckCurrent() so that it doesn't
    create an "inconsistent" JniEnvironment w/ a null JavaVM.