[Java.Interop] Move JavaVM.LocalReferenceCount to JniEnvironment.
The local reference count -- at least on Android -- is tied to the JNI
Environment (JNIEnv*), not the VM itself. (One immediate implication
of this is that you CANNOT share local references between threads; you
need to create a global reference first, and share the global
reference between threads.)

Since the local reference count is a per-JniEnvironment (and per-
thread) value, the local reference count should exposed as
JniEnvironment.LocalReferenceCount, not JavaVM.LocalReferenceCount.

However, since JniEnvironment cannot be subclassed, preserve and
update the JavaVM.LogCreateLocalRef() and JavaVM.LogDestroyLocalRef()
methods to take the JniEnvironmentSafeHandle for the JNIEnv* that the
local ref is being created for.