[Java.Interop] Rename JavaArrayElements to JniArrayElements.
As mentioned in README.md (8d9b11e9) and 429cdc38, the "Java" prefix
should be used for types which can participate in cross-VM object
graphs.

JavaArrayElements was _not_ such a class; it wraps a raw pointer to an
unmanaged primitive array. As such, it should not have a "Java" prefix.

Rename JavaArrayElements to JniArrayElements, and rename
JavaArrayElementsReleaseMode to JniArrayElementsReleaseMode.