[Java.Interop] Rename the JNIEnv semantic wrappers.
The concern is one of "namespace pollution"- there are 11 "semantic
wrappers" for the Java JNIEnv structure: JniActivator, JniArrays,
JniErrors, JniHandles, JniIO, JniMembers, JniMonitors, JniStrings,
JniTypes, JniJavaVMs, JniVersions.

If all of these were to be made public, that would be an additional 11
public types that would show up in IDE completion for "Jni", most of
which will not be relevent. (This is why most of these types aren't
public, but if (when?) they were public...)

Furthermore, all of these types are tied to JniEnvironment.Current,
and that's not immediately obvious (without documentation).

Consequently, "rename" them, removing the "Jni" prefix and nesting
them within JniEnvironment. Thus, JniActivator becomes
JniEnvironment.Activator, JniArrays becomes JniEnvironment.Arrays,
etc.

This has the side effect of potentially making these types harder to
use (which is why the Framework Design Guidelines suggest avoiding
them [0]), which is actually beneficial here: These types should not
be used unless there's no better wrapper. For example,
JniEnvironment.Arrays shouldn't be used directly; that's what
JavaArray/etc.  are for.

[0]: http://msdn.microsoft.com/en-us/library/ms229027(v=vs.110).aspx