[Java.Interop.Export] Marshal IJavaObject return values.
The primary question about [Export] methods returning IJavaObject is
what should be returned. The "obvious" thing to do is to have the
JNI marshaling method "just" return IJavaObject.SafeHandle, i.e. the
JNI marshal method would resemble:

	JniReferenceSafeHandle __marshal__ (IntPtr jnienv, IntPtr context /* ... */)
	{
		...
		IJavaObject ret = CallSomeMethod();
		return ret == null || ret.SafeHandle.IsInvalid ? null : ret.SafeHandle;
	}

(This is in fact what I originally tried to do just to have something
for test purposes, but Mono would keep complaining that the IL was
invalid...)

The problem with the obvious solution is that eventually we'll need to
"deep marshal" collections (not currently supported, but eventually!),
e.g. we have an [Export] method which returns an array:

	[Export ("funcInt32Array", Signature="()[I")]
	static int[] FuncInt32Array ()
	{
		return new[]{1,2,3};
	}

If we stick with the "obvious" solution, we end up "burning" a GREF:

	JniReferenceSafeHandle __marshal__ (IntPtr jnienv, IntPtr context /* ... */)
	{
		...
		int[] r = FuncInt32Array ();
		var __r = new JavaInt32Array (r.Length);
		__r.CopyFrom (r, 0, 0, r.Length);
		return __r.SafeHandle;
	}

`__r` holds a GREF, so we'd need (at best!) a full GC in order to
collect the JavaInt32Array instance. (At worst, it's a memory leak.)

The fix for this corner case is to always return a JNI Local Reference
from JNI marshal methods. Xamarin.Android would effectively do:

	IntPtr __marshal__ (IntPtr jnienv, IntPtr context /* ... */)
	{
		...
		int[] r = FuncInt32Array ();
		return JavaArray<int>.ToLocalJniHandle (r);
	}

The benefit here is that JavaArray<T>.ToLocalJniHandle() can use the
GREF-using JavaArray<T> type, then create a JNI local reference over
the GREF, then dispose of the JavaArray<T>. The JNI local reference
will keep the Java instance alive, and we no longer leak a GREF.

Clearly then, the Java.Interop.Export approach should be to likewise
return a JniLocalReference, right?

	JniLocalReference __marshal__ (IntPtr jnienv, IntPtr context /* ... */)
	{
		...
		int[] r = FuncInt32Array ();
		using (var __r = new JavaInt32Array (r.Length)) {
			__r.CopyFrom (r, 0, 0, r.Length);
			return __r.SafeHandle.NewLocalRef ();
		}
	}

Wrong. The problem with using JniLocalReference is that it's a
GC-tracked value, meaning at some point the GC will attempt to dispose
of the instance, which will attempt to free the JNI local ref, which
will fail horribly because it shouldn't be freed! (We're returning
to JNI, passing ownership of the JNI local reference to JNI.)

Thus what we need is something that won't be collected by the GC.

We could use JniInvocationHandle, but:

 1. That would still require allocating a GC object for little gain;
 2. We would still need a way of getting a JNI local reference w/o
    using a JniLocalReference instance.

Which brings us to the new JniEnvironment.Handles.NewReturnToJniRef()
method, which is a wrapper/rewrite of JNIEnv::NewLocalRef(), _except_
instead of returning a JniLocalReference it returns an IntPtr, with
the semantic that it is to ONLY be used when the created value will be
a return value returned to JNI. (There is no DeleteReturnToJniRef(),
deliberately so.)

This allows our hypothetical collection marshaling code to create a
"ReturnToJniRef" (JNI local reference) without maintaining a GREF:

	IntPtr __marshal__ (IntPtr jnienv, IntPtr context /* ... */)
	{
		...
		int[] r = FuncInt32Array ();
		using (var __r = new JavaInt32Array (r.Length)) {
			__r.CopyFrom (r, 0, 0, r.Length);
			return JniEnvironment.Handles.NewReturnToJniRef (__r.SafeHandle);
		}
	}

Note: This may all be a horrible mistake, as we'll need to be able to
deep-marshal collections inside collections inside... (e.g.
`int[][][]` marshaling shouldn't require any GREFs once complete), and
I'm not quite sure how the intermediate marshals will be held.

Related: improve the tests so that instead of using JNI to invoke the
`native` methods, we add Java-side testStaticMethods() and
testMethods() methods, which (1) invoke the `native` methods, and (2)
can check return values to ensure that our marshaling is correct.

This also found an error in the tests: ExportType.actionInt32String()
was declared as static in Java but instance in C#, resulting in a
runtime error when attempting to execute. Doh!