[Java.Interop-PerformanceTests] Test JniMethodID caching ideas.1
With the realization that how Xamarin.Android has been doing "base"
method invocation is ALL WRONG (see 634a9aaf), I have to wonder if the
Xamarin.Android approach for caching JniMethodIDs is correct as well.

Thus the TimingTests.MethodLookupTiming() timing, which invokes
Object.toString() 1 million (pinky-in-mouth) times using different
JniInstanceMethodID cache strategies to see how they compare.

The strategies are:

 1. "Traditional": Have a static field which holds the
    JniInstanceMethodID, which is looked up via
    JniTypeInfo.GetCachedInstanceMethod().

 2. "No caching": Don't cace the JniInstanceMethodID, just look it up
    new each time we need it.

 3. "Cache in a Dictionary w/ lock": use a
    Dictionary<string, JniInstanceMethodID> and a C# `lock` on the
    dictionary to lookup a cached JniInstanceMethodID value.

 4. "Cache in a ConcurrentDictionary": lock-free alternative to (3).

While (1) can be "made to work", if it's no longer viable to use
JNIEnv::CallNonVirtual*Method() with a "base" jmethodID, I don't see a
whole lot of point of having a bazillion static members to cache
JniMethodID values for each method. We can thus shrink our memory
overheads (static fields) by caching only the members we need.
(In my brain, anyway; it's possible that dictionary overheads would be
larger than the "unused" static fields, and when you have a linker --
as is the case for Xamarin.Android -- the "unused" fields will
actually be used.)

So how do they stack up?

	Method Lookup + Invoke Timing:
	           Traditional: 00:00:01.5189943
	            No caching: 00:00:03.5178602
	          Dict w/ lock: 00:00:01.5425098
	        ConcurrentDict: 00:00:04.9707850

(Times of course vary from run to run, but this is fairly
representative.)

The best time is the Traditional approach (which is encouraging -- we
never actually timed/profiled the approach, just reasoned it out).

Second best is the Dictionary w/ `lock`, which isn't significantly
worse. This is also encouraging.

No caching is terrible, taking ~2.3x as long to complete as the
traditional approach. This isn't surprising, as invocation of the
JniInstanceMethodID requires an "additional" JNI transition to lookup
the jmethodID value, ~doubling the "overhead" of invocation.
This is clearly a non-starter.

ConcurrentDictionary is disappointing: it performs worse than no
caching at all, taking 3.2+x as long to execute.

I had no idea Mono's ConcurrentDictionary was so slow; I'll need to
investigate its use in JavaVM (JavaVM.JavaVMs,
JavaVM.TrackedInstances), as it seems plausible that for our typical
use cases a Dictionary with a lock may be faster.

Based on these results, I plan on prorotyping a Dictioary-based method
caching mechanism to see how that plays out.