[Java.Interop] Fix JniEnvironment.Handles.NewReturnToJniRef().
The problem is that JniEnvironment.Handles.NewReturnToJniRef() would
occasionally return an invalid handle, causing the process to abort:

	Invalid memory access of location 0x14 eip=0x1b268a
	Stacktrace:

	Native stacktrace:

I don't know why the process is aborting, but NewReturnToJniRef() was
clearly doing something "wrong" (or my delegate signature is wrong,
but it looked right to me...).

Add a new JniEnvironmentTests.Handles_NewReturnToJniRef() test which
uses System.identityHashCode() (expose as
JniEnvironment.Handles.GetIdentityHashCode()) to see if the handle
created by NewReturnToJniRef() corresponds to the input value (which
is what triggers the above crash, but the test fails _before_ the
crash, so it's still a useful test).