[Java.Interop] Rework method invocation API.
As per b63260db, getting method invocation right can be tricky.

Making it efficient moreso (b63260db doesn't cache jmethodIDs as much
as I'd like).

Making a sane invocation API (for use by humans and generators alike)
that doesn't result in massive code bloat, while making things
reasonably efficient? Uh...

Now we can observe the timing results of 910342b1, which shows that
using locking on a Dictionary to cache jmethodID values is not much
worse than using e.g. JniType.GetCachedInstanceMethod(). This
observation allows us to do the following:

 1. Add a JniPeerMembers type.
 2. Add an IJavaObject.JniMembers property (TODO: rename).

JniPeerMembers has three responsibilities:

 1. It "merges" ye olde Xamarin.Android Object.ThresholdType and
    Object.ThresholdClass properties.
    (JniPeerMembers.ManagedPeerType == Object.ThresholdType,
    JniPeerMembers.JniPeerType == Object.ThresholdClass.)

 2. It provides jmethodID/jfieldID caching services.

 3. It provides invocation support.

(1) is required so that (3) can Do The Right Thing when it comes to
virtual vs. non-virtual method dispatch (to determine whether virtual
or non-virtual dispatch is required).

(2) is to help solve the efficiency issues posed by b63260db.

(3) is to make things easier in the binding code -- see the diffs in
JavaObject and the CallNonvirtual types to see how much shorter the
binding methods become [0]!. It also serves to "unify" the
virtual-vs-nonvirtual invocation logic, making it harder to screw up.

The "cute" part is that non-virtual dispatch involves _two_
JniPeerMembers instances: the JniPeerMembers for the declaring type
(e.g. CallNonvirtualBase), and the JniPeerMembers instance for the
"threshold" type (the Java framework peer and corresponding managed
peer types). This allows us to use the correct jmethodID and jclass
for the JNIEnv::CallNonvirtual*Method() AND CACHE the jmethodID of the
threshold class (CallNonvirtualDerived.method(), in this case).

[0]:  Alas, the binding support is WAY incomplete: only methods which
      take no parameters and have an Object, IJavaObject, or int
      return type are supported. This support will need to be vastly
      improved in order to be usable, and will likely require a code
      generator.