[Java.Interop] Simplify JavaObject subclass creation.
Creating instances of types which have a default constructor shouldn't
be difficult, yet it was: each type would re-declare _NewObject(),
which would create a JniType instance for the corresponding Java Peer
type, lookup the default constructor, and create a new instance using
the default constructor.

The reason all this pain was necessary was twofold:

 1. To permit caching the jmethodID for the constructor.
 2. When JavaObject was written, there wasn't a way to lookup the
    Java Peer type for a Managed Peer type.

(2) was solved in e6d86e3f with JavaVM.GetJniTypeInfoForType(Type),
which allows looking up the Java Peer type for a Managed Peer type.
(1) isn't strictly necessary, as in Xamarin.Android there are many
circumstances in which the jmethodID isn't cached (largely when
constructing instances of User Peer types).

We have the plumbing: use it!

The JavaObject default constructor will now use
JavaVM.GetJniTypeInfoForType(Type) to determine the Java Peer type for
the current type. If there is no such type, a NotSupportedException is
thrown; if there IS such a type, it is used for instance construction.