[Java.Interop-Tests] Add more JavaObject assertions.
ermau requested awhile ago that ~all member invocations on disposed
objects should throw ObjectDisposedException. (This still isn't the
case as of Xamarin.Android 4.12.0 -- attempting to invoke a method on
a disposed instance will result in an ArgumentException from JNIEnv.)

This is truly sane and correct behavior: ensure that JavaObject throws
ObjectDisposedException when the instance has been disposed.

Note: not all members should throw, in particular
JavaObject.SafeHandle and JavaObject.JniPeerMembers. The rationale for
this is twofold:

 1. JniPeerMembers is invoked virtually from the constructor, before
    SafeHandle has been set, and thus the object is in an "invalid"
    state but the JniPeerMembers property MUST be used.

 2. It is useful to be able to access/compare SafeHandle on disposed
    instances, e.g. to workaround/diagnose GC bugs. If SafeHandle
    threw, these circumstances would be hindered.