[TestJVM, Java.Interop-Tests] Add & Sanitize LREF logging.
Add logging for JNI local references to JavaVM if _JI_LOG contains an
"lref" or "lref=<PATH>" value (writing output to <PATH> if specified).

Review the resulting JNI local reference logging, and notice that the
local reference counts are going *negative*, and not just negative,
but _hundreds_ negative.

Review of the logs raises two things (thus necessitating a log format
change from Xamarin.Android):

 1. Since JNI local references are tied to their corresponding JNIEnv,
    and because JNIEnv is a thread-local value, JNI local references
    are _also_ thread-local values (which is why we moved
    JavaVM.LocalReferenceCount to JniEnvironment.LocalReferenceCount
    back in d45dcf56).

    Properly understanding this requires that the log messages specify
    which thread performed the reference new and delete.
    (This should likewise be done for other references.)

 2. By adding the thread name+ID to the log message (1), I found the
    reason for the hugely negative JNI local reference count: the
    JNI local reference was created on one thread, but was being
    destroyed on the Finalizer thread. Since reference destruction
    decrements the reference count of the JniEnvironment on the
    _current_ thread, the result was lots of increases in one thread,
    but the corresponding decreasee was on another thread.

I "fixed" the issue by fixing the unit tests to ensure that values
were disposed on the thread that created them instead of relying on
the finalizer.

Now the only JNI local references which are disposed on the Finalizer
thread are values tied to the JniEnvironment class itself, so it's
reasonable for them to only get cleaned up by the finalizer.

Note/TODO: "Log review" still seems like a hack. Perhaps
JniLocalReference should also contain a reference the JniEnvironment
it was created from, allowing it to decrement the correct
JniEnvironment value when it's finalized?

However, finalizing JniLocalReferences from the finalizer thread is
still an open question; see e3c48e7b. I have a suspicion that
ART/Dalvik will abort the process in the event of cross-thread JNI
local reference destruction; this still needs to be tested.