[Java.Interop] Add JniEnvironment.Types.GetJniTypeNameFromInstance().
This is the equivalent of Xamarin.Android's
JNIEnv.GetClassNameFromInstance(). It provides the JNI type name for
an instance of a class, e.g. an instance of java.lang.Object would
provide a JNI Type name of "java/lang/Object".

(Note that java.lang.Class.getName() returns "dotted' names, so we
replace '.' with '/' to obtain the JNI name. Interestingly, for nested
types Class.getName() uses '$' as the typename separator, e.g.
`java.lang.Thread.State.class.getName()` is "java.lang.Thread$State".
Since '$' is the nested type seprator in JNI as well, this is kept.)