[Java.Interop] Add JavaObjectExtensions.GetJniTypeName().
Add a JavaObjectExtensions.GetJniTypeName() extension on IJavaObject.
This is a small wrapper around
JniEnvironment.Types.GetJniTypeNameFromInstance(), intended to be more
discoverable and useful (i.e. it'll show up in IDE code completion on
IJavaObject, as opposed to "knowing" about JniEnvironment.Types).

Why an extension method? Becuase there's no need for this method to
exist on IJavaObject: the implementation would be the same for all
current (and possible future) IJavaObject implementations. I thus
don't see the need to "burden" all the implementors with this method.

(It'll also be handy when JavaCast<T>() is eventually provided, which
likewise should be an extension method and not part of IJavaObject
itself.)