[Java.Interop] JNI local reference cleanup.
Reviewing the JNI local reference logging, I see that that the
finalizer thread is getting to -61 JNI local references, many of which
are "bugs"/"leaks" that the GC is helpfully cleaning up.

Except we don't want the GC to be cleaning up all our messes...

Initial pass:

  * Fix JniMarshal.GetValue<T>() to dispose of the reference
    when the value is a registered JavaProxyObject (which was leaked);
  * JavaException.Dispose() should Dispose() the InnerException,
    if possible;
  * CollectionContract<T>.Contains() didn't dispose of created values.
  * JavaVMTest.GetObject_ReturnsAlias() didn't dispose the alias.

With the above fixes, we're down to -37 JNI local references on the
finalizer thread. (Still not wonderful, but an improvement.)

Observation: Many of these "leaks" are due to JavaVM.GetObject()
calls in which the returned value isn't disposed...because it can't
be. (Not reliably, anyway, as the value returned may be a registered
value, and always disposing of that would be Bad.) For example, see
the TestType marshal methods.