[Java.Interop] Add IJavaObject.DisposeUnlessRegistered().
This tackles the observation from 2ab6e735:

> Many of these "leaks" are due to JavaVM.GetObject() calls in which
> the returned value isn't disposed...because it can't be.

The problem is marshal methods, e.g. the TestType.GetInt32Value()
marshal method, paraphrased as:

	// BROKEN: "leaks" a JNI local reference
	int GetInt32Value_Marshaler (IntPtr jnienv, IntPtr n_self)
	{
		var self = JniEnvironment.Current.JavaVM.GetObject<TestType>(n_self);
		return self.GetInt32Value
	}

The above is clearly a possible JNI local reference "leak": if
`n_self` doesn't belong to a registered instance, then
JavaVM.GetObject<T>() MUST create a *new* instance (484b44e7), which
will contain a JNI local reference.

How can we free this JNI local reference? We _cannot_ Dispose() it:

	// WRONG
	int GetInt32Value_Marshaler (IntPtr jnienv, IntPtr n_self)
	{
		using (var self = JniEnvironment.Current.JavaVM.GetObject<TestType>(n_self))
			return self.GetInt32Value
	}

The above is invalid because if n_self DOES correspond to a
registered instance, then the implicit Dispose() will NUKE that
registration. This is clearly undesirable.

There are three possible solutions here:

 1. Move IJavaObjectEx.Registered to IJavaObject.
 2. Change the return type of JavaVM.GetObject().
 3. Add IJavaObject.DisposeUnlessRegistered();

We could move IJavaObjectEx.Registered to IJavaObject. This is
undesirable because it makes the "state" of IJavaObject more explicit,
thus complicating what (potentially) user-written code must do:

	// THEORETICAL FIX: NOT DONE
	int GetInt32Value_Marshaler (IntPtr jnienv, IntPtr n_self)
	{
		var self = JniEnvironment.Current.JavaVM.GetObject<TestType>(n_self);
		try {
			return self.GetInt32Value
		} finally {
			if (!self.Registered)
				self.Dispose ();
		}
	}

Alternatively, we could change the return type of JavaVM.GetObject()
to something for which the Dispose() method Does The Right Thing:

	struct JniRegistrationEntry<T> : IDisposable
		where T : IJavaObject
	{
		public T Value {get;}
		public void Dispose ()
		{
			var ex = Value as IJavaObjectEx;
			if (ex != null && !ex.Registered)
				Value.Dispose ();
		}
	}

	partial class JavaVM {
		public JniRegistrationEntry<T> GetObject<T>(IntPtr handle);
	}

The problem with the above is it complicates all uses of
JavaVM.GetObject(): instead of an IJavaObject (or instance of the
desired type), it instead returns a "wrapper" which must be
"dereferenced" in order to obtain the desired instance.

I don't see this as being a net gain by any actual measure:

	// THEORETICAL FIX: NOT DONE
	int GetInt32Value_Marshaler (IntPtr jnienv, IntPtr n_self)
	{
		JniRegistrationEntry<TestType> self = JniEnvironment.Current.JavaVM.GetObject<TestType>(n_self);
		using (self) {
			// Note the extra `.Value` to "dereference" the entry
			return self.Value.GetInt32Value ();
		}
	}

Adding IJavaObject.DisposeUnlessRegistered() is the chosen solution:
it allows the IJavaObject implementation to determine whether it
should be disposed, based on whether IJavaObject.RegisterWithVM() has
been called on the instance.

  * If IJavaObject.RegisterWithVM() has NOT been called, then
    IJavaObject.DisposeUnlessRegistered() is synonomous with
    IJavaObject.Dispose().
  * If IJavaObject.RegisterWithVM() HAS been called, then
    IJavaObject.DisposeUnlessRegistered() is a no-op.

This allows our marshal method to become:

	// FINAL MARSHAL METHOD
	int GetInt32Value_Marshaler (IntPtr jnienv, IntPtr n_self)
	{
		var self = JniEnvironment.Current.JavaVM.GetObject<TestType>(n_self);
		try {
			return self.GetInt32Value
		} finally {
			self.DisposeUnlessRegistered();
		}
	}

Unfortunately the above can't use a `using` block -- changing
JavaVM.GetObject() would allow that -- but it seems to be the best
solution given the constraints.

With these changes in place, we go from -20 JNI local references on
the Finalizer thread to -2 JNI local references; much better.