[Java.Interop] Unify names & signatures in JniPeerMembers.
(Fun with embedded NULs!)

The problem is that JniPeerMembers.CallInstanceInt32Method() and
related are ugly:

	_members.CallInstanceInt32Method ("hashCode", "()I", "hashCode()I", this);

Why was it designed like that in ce9c1b2e? To reduce allocations.

On the "hot path" for methods that have already been looked up, the
intention was that the 3rd parameter -- "hashCode()I" -- would be used
as the JniPeerMembers.InstanceMethods key, and I wanted this to be a
constant so that we could perform the InstanceMethods lookup without
needing any allocations.

Meanwhile we also needed separate 'name' and 'signature' values so
that we could fallback to perform the JniType.GetInstanceMethod()
lookup to fill the cache. This resulted in three parameters.

However, this is very, very, ugly.

Rework this so that instead of three parameters we use an
"encodedMember", a single string parameter with an embedded ASCII
NULL -- \u0000 -- to separate the member name from the signature. We
can then "split up" the encodedMember into the separate member name
and signature in the cache-miss path.

(Yes, there will be required memory allocations now in the cache miss
path, but ideally there won't be that many cache misses once the app
is actually running.)

This simplifies the above call to:

	_members.CallInstanceInt32Method ("hashCode\u0000()I", this)

That is a tad odd to look at and write manually, but long term we'll
have a generator emitting these signatures, and it's not _that_ bad to
hand-write either. It's certainly better than the previous approach!