[Java.Interop] Prototype JniPeerInstanceMethods.CallInt32Method<T>().
(Time for an epic rant on the bigger picture... ;-)

What's going on here? Why do we need recursive array marshaling
(77c1a4cf)? Why a "threat" of ~160 methods in
JniPeerMembers.InstanceMethods (a088ebcb)?

The idea is simple: reduce (generated) "code bloat" and make it easier
to write hand-bound methods by making the parameter marshaling
infrastructure _generic_.

For example, consider the new TestType.updateInt32ArrayArray()
Java method:

	// Java
	class TestType {
			public int updateInt32ArrayArray (int[][][] array);
	}

To bind this in Xamarin.Android, the generator would emit:

	static IntPtr id_update;
	public int updateInt32ArrayArray (int[][][] array)
	{
		// (1): jmethodID lookup
		if (id_update == IntPtr.Zero)
			id_update = JNIEnv.GetMethodID (class_ref, "updateInt32ArrayArray", "[[[I");

		// (2): parameter marshaling
		IntPtr native_array = JNIEnv.NewArray (array);

		// (3): method invocation
		int __ret;
		if (GetType () == ThresholdType)
			__ret = JNIEnv.CallIntMethod (Handle, id_update, new JValue (native_array));
		else
			__ret = JNIEnv.CallNonvirtualIntMethod (Handle,
					ThresholdClass,
					JNIEnv.GetMethodID (ThresholdClass, "updateInt32ArrayArray", "[[[I"),
					new JValue (native_array));

		// (4): (optional) copy array contents from Java to managed code.
		if (array != null) {
			JNIEnv.CopyArray (native_array, array);
			JNIEnv.DeleteLocalRef (native_array);
		}
	}

Compare to what is needed in Java.Interop with this commit:

	public int UpdateInt32ArrayArray (int[][][] array)
	{
		return _members.InstanceMethods.CallInt32Method ("updateInt32ArrayArray\u0000([[[I)I", this, value);
	}

Those two bodies are actually equivalent: both perform the
JNIEnv::GetMethodID() lookup and caching, both dispatch to
JNIEnv::CallIntMethod() or JNIEnv::CallNonvirtualIntMethod() depending
on the runtime type, both completely marshal the array parameter into
Java land and marshal the result back out into Managed land.

They do the same things. HOW they do them is completely different.

OK, but WHY? Hand-binding is still occasionally needed, and the
emitted code is quite bloated, so that's for starters.

The Big UNKNOWN is what the PERFORMANCE implications of relying upon a
generic marshaling infrastructure will be. With the generated code,
marshaling is as efficient as it can be, in that there isn't any ADDED
overhead in marshaling the values before performing the call.

Compare to the new approach, which has AT LEAST one
JavaVM.GetJniMarshalInfoForType() virtual method invocation PER
PARAMETER, several structs which are copied, and delegate invocations
to perform marshaling (both to- and from- JNI) and cleanup. Nearly all
of that overhead is avoided in the Xamarin.Android implementation.

Which is partially why to implement it: to FIND OUT what the
performance implications actually are. (I don't have those timing
results yet, and won't until I flesh this out, but the overall shape
of generic parameter marshaling should now be apparent...)

Implementation notes:

JavaObjectArray<T>.forMarshalCollection is present so that we can
ensure that JNI local references are not "leaked" when grabbing
element values which are themselves temporary objects as part of the
"copy back to managed code" logic.

JniMarshalInfo.MarshalFromJni has been renamed to
JniMarshalInfo.GetValueFromJni. JniMarshalInfo.MarshalToJni has been
renamed to JniMarshalInfo.CreateLocaRef. The reason for this is that
we now need many different ways of marshaling to JNI, and now provide
*three*

  * JniMarshalInfo.CreateJValue(): Needed for marshaling builtin value
    types, e.g. `int`, which do not use JNI handles.
  * JniMarshalInfo.CreateLocalRef(): When there is no need to "copy
    back" the value. This is used by normal JavaObject marshaling.
  * JniMarshalInfo.CreateMarshalCollection(): When an intermediate
    "collection" is created which must be "copied back" into Managed
    code. This is primarily for arrays.

Furthermore, if JniMarshalInfo.CreateMarshalCollection is used,
JniMarshalInfo.CleanupMarshalCollection can be provided to actually
perform the "copy back" behavior and delete the intermediary.

JniMarshalInfo is a struct type so that the delegates can be copied
via blitting and to reduce the number of distinct types that must be
mentally tracked.

Note: JniMarshalInfo being a struct is likely a mistake, given that --
historically! (new tests are needed) -- delegate invocation is
generally slower than virtual method invocation. This should be timed,
and possibly replaced with a normal class and virtual methods.