[Java.Interop-PerformanceTests] Measure generic marshaling overheads.
Commit 09bc95ce asks:
> The Big UNKNOWN is what the PERFORMANCE implications of relying upon a
> generic marshaling infrastructure will be. With the generated code,
> marshaling is as efficient as it can be, in that there isn't any ADDED
> overhead in marshaling the values before performing the call.

Let's measure that, for two different virtual Java methods: one with a
builtin type (int), and one for an array (int[][][]).

Note: this isn't extensive: we're assuming a virtual method invocation
that could be overridden, e.g. we need the
`GetType() == ManagedPeerType` check (a `final` method could skip
this), and no classes overridden.

Results:

	Generic Marshaling Overhead: (I)I
	         Native Marshaling: 00:00:00.1165644
	        Generic Marshaling: 00:00:00.1475230
	Generic Marshaling Overhead: ([[[I)I
	         Native Marshaling: 00:00:02.1584246
	        Generic Marshaling: 00:00:02.6341127

The (int) marshaling overheads results in ~26% longer execution time.
The (int[][][]) overhead results in ~22% longer execution time.

This implies a current generic marshaling upper bound of ~27%
depending on the type being marshaled for a Java method which does
nothing. (Overhead will of course "improve" as the Java method does
more work/Java exceution time increases.)

This isn't wonderful, but it's not as bad as I feared.