[TestJVM] Reformat LREF/GREF messages.
Change the JNI local reference messages so that the thread name is at
the _end_ of the message, not the start. This improves compatibility
with Xamarin.Android, which should allow existing logcat parse tools
to be used (or with minimal change).

Change the JNI global reference messages so taht the thread name is at
the end of the message (as opposed to not provided at all). Thread
information can be extremely handy. (It's likewise at the end so that
the overall format is ~identical with Xamarin.Android.)