[Java.Interop] Provide Debug.Assert() messages.
When Debug.Assert() is "raised" (fired? invoked?), it prints the
message text and a stack trace of where the assert failed.

Unfortunately, I wasn't providing any message text for most of the
Debug.Assert() messages, meaning there was no useful message stating
_why_ the assert was failing. Oops.

Provide a message so that the asserts are useful.

This in turn allowed me to determine _why_ the asserts were failing:
JavaVMTest.AssertGetJniMarshalInfoForPrimitiveType() was passing
`null` as the targetType parameter -- which should be fine -- but the
builtin's GetValueFromJni() methods were asserting that it not be
null. Fix the builtin's GetValueFromJni() assert to permit a null
targetType, which means "use your default targetType."