[Java.Interop] Exception propogation!
In commit 782fa02e:

> Option 2: D doesn't have a (JniReferenceSafeHandle, JniHandleOwnership)
> constructor.
>
> 	  C#: JavaVM.GetObject() throws an exception
> 	**BOOM**
>
> Becuase we don't have correct exception propogation implemented yet,
> this results in corrupting OpenJDK (we unwound native Java stack
> frames!), and the process later aborts:
>
>	malloc: *** error for object 0x79d8c248: Non-aligned pointer being freed
>	*** set a breakpoint in malloc_error_break to debug
>	Stacktrace:

Clearly memory corruption is BAD. (VERY VERY BAD.)

Crashes that happen nowhere near the cause of the crash are equally
bad (as they cause/contribute to memory corruption).

The lynchpin to exception propogation is the new
JniMarshalMethod.CreateMarshalMethodExpression() method, which uses
System.Linq.Expressions to "wrap" a JNI marshal method invocation
within a try/catch block. It also now performs the
JniEnvironment.CheckEnvironment() check.

JniType.RegisterNativeMethods() has been updated to "wrap" the
delegates to register with JNI using
JniMarshalMethod.CreateMarshalMethodExpression().

The result ("flattening" a bit) is that JNI marshal methods which
previously were this:

	static void JniMarshalMethod (IntPtr jnienv, IntPtr n_self)
	{
		JniEnvironment.CheckEnvironment (jnienv);

		var self = JniEnvironment.Current.JavaVM.GetObject<TestType> (n_self);
		try {
			self.MethodThrows ();
		} finally {
			self.DisposeUnlessRegistered ();
		}
	}

now (effectively) become this:

	static void JniMarshalMethod (IntPtr jnienv, IntPtr n_self)
	{
		JniEnvironment.CheckEnvironment (jnienv);

		try {
			var self = JniEnvironment.Current.JavaVM.GetObject<TestType> (n_self);
			try {
				self.MethodThrows ();
			} finally {
				self.DisposeUnlessRegistered ();
			}
		} catch (Exception e) {
			JniEnvironment.Errors.Throw (e);
		}
	}

(Note that it's not _actually_ flattened; see JniMarshalMethodTests
for details.)

Assessories to this are the new JniEnvironment.Errors.Throw(Exception)
method and the JavaProxyThrowable type. JavaProxyThrowable is a
JavaException subclass which holds an Exception. It can be raised in
in managed code via
JniEnvironment.Errors.Throw(JniReferenceSafeHandle), will propogate
"properly" through Java code (it's a java.lang.Throwable), allowing
Java-side finalizers and catch blocks to behave properly, and if the
Java code doesn't catch it the exception will re-enter managed code.

At which point more magic happens: since the JavaProxyThrowable is
registered, JavaVM.GetExceptionForThrowable() will find it, and is
special-cased so that the wrapped exception is re-thrown.

Thus, if we have a Managed > Java > Managed stack, and the rightmost
method throws a managed exception, the Java code will properly have a
chance to unwind, and the leftmost Managed frame will be able to catch
the originally thrown exception.

Magic! No memory corruption! Proper Behavior! (Yay!)