[Java.Interop] JavaVM.GetObject() returns best matching type.
Now that we have JNI type reference to System.Type mapping support
(84c39a2f), we can now implement "best matching type" support.

If an un-wrapped Java instance enters managed code, a wrapper must be
constructed to hold the wrapper. What type should be used?

An answer is to use just the specified type, e.g. either the
JavaVM.GetObject<T>() type parameter, or the targetType parameter.

The problem with this is (not yet supported) Java-side activation,
in conjunction with base method invocations from constructors
(782fa02e). In this scenario, there would be a Managed peer + Java
peer, and a base class constructor of the Java peer would invoke a
method that is overridden by the Managed peer (using
Xamarin.Android methods):

	// Java: Base constructor() invokes virtual method
	public class Base {
		public Base () {
			method ();
		}
		public void method() {
		}
	}

	// Android Callable Wrapper/Java Callable Wrapper
	public class NativePeer extends Base {
		public NativePeer () {
			super ();
			if (getClass () == MyAsyncTask.class) {
				mono.android.TypeManager.Activate (FullyQualifiedAssemblyName, JniSignature, this, new java.lang.Object[] { /* parameters */ });
			 }
		}
		@Override
		public native void method();
	}

Connected to this, there would be the binding of Base and the
ManagedPeer that corresponds to NativePeer:

	// C#: Base binding
	public class Base : JavaObject {
		public virtual void Method () {/* ... */}

		// JNI marshal method for Base.method() overrides
		static void JniMarshalMethod (IntPtr env, IntPtr n_self)
		{
			var self = JavaVM.Current.GetObject<Base>(n_self);
			n_self.Method ();
		}
	}

	public class ManagedPeer : Base {
		public override void Method () {/* ... */ }
	}

Note that in "normal" Xaamrin.Android, there would be a JNI marshal
method declared in the base class, not in each derived class (as is
currently done in CallVirtualFromConstructorDerived).

With that in mind, we now have a problem when instances are first
constructed from Java, not managed code:

	Java: `new NativePeer()`
	Java: Base.<init>();
	Java: NativePeer.method() [virtual override]
	  C#: Base.JniMarshalMethod()
	  C#: JavaVM.GetObject<Base>()

We now hit (at least) one conundrum: What type should
Base.JniMarshalMethod() create? It calls JavaVM.GetObject<Base>(), but
the _runtime type_ of the instance passed will be NativePeer, which is
bound as ManagedPeer.

The Xamarin.Android answer -- and thus the correct answer ;-) [0] --
is for JavaVM.GetObject() to return the "best matching" type that it
possibly can. Thus, JavaVM.GetObject<Base>() would create an instance
of ManagedPeer -- a type which inherits from Base.

Update JavaVM.GetObject() to implement this "best match" logic.

Test note: There is currently no logic (and there likely will be no
logic) to automagically map a Java type to a C# type; this is the role
of the virtual JavaVM.GetTypeForJniSimplifiedTypeReference() method.

TestJVM now overrides GetTypeForJniSimplifiedTypeReference(), and
takes a Dictionary<string, Type> constructor parameter which contains
the JNI type reference :: System.Type mapping. This in turn requires
updating the TestJVM constructor, and updating some existing tests.

[0]: Note: Xamarin.Android is not always the correct answer. ;-)