[Java.Interop] Zero-GREF marshaling for builtins.
Consider the innocuous JavaObjectArray_object_ContractTest, which
effectively does:

	var array = new JavaObjectArray<object>(3);
	array [0] = new object ();
	array [1] = 42;
	array [2] = "C";

Reasonably straightforward, but there's a hidden GREF cost: because we
have a JavaObjectArray<object>, we use
JavaVM.GetJniMarshalInfoForType(typeof(object)), which doens't provide
any marshaling info, so we use the fallback JavaProxyObject.GetProxy()
method to "convert" e.g. 42 into an IJavaObject, and the proxy must
use a GREF.

Which is fine for array[0], but both array[1] and array[2] have proper
marshaling support! Yet we'd still take out a GREF for the int and the
string constant!

Blech!

Improve JniMarshal.CreateLocalRef<T>() and JniMarshal.GetValue<T>() so
that they take the runtime type into consideration, ahead of `T`.

Within JniMarshal.CreateLocalRef<T>(), this involves using
JavaVM.GetJniMarshalInfoForType(value.GetType()) (if value is not
null), falling back to using
JavaVM.GetJniMarshalInfoForType(typeof(T)) if the value is either null
or no marshaler was found for value.GetType().

Similarly, within JniMarshal.GetValue<T>(), before using
JavaVM.GetObject() we first lookup the JNI type of the target value,
and do the JavaVM.GetTypeForJniTypeRefererence() /
JavaVM.GetJniMarshalInfoForType() dance to find an optimized
marshaler.

By doing this, we can ensure that no additional GREFs are used to
store 42 or "C" within the JavaObjectArray<object>.